// chpqa-app/src/app/api-services/chpqa-api/generated/model/replyMeter.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OutputUnit } from './outputUnit';
import { ReplyEquipmentType } from './replyEquipmentType';
import { ReplyMeterFile } from './replyMeterFile';
import { ReplyEquipmentSubType } from './replyEquipmentSubType';


export interface ReplyMeter { 
    id?: string | null;
    name?: string | null;
    tagNumber?: string | null;
    equipmentType?: ReplyEquipmentType;
    equipmentSubType?: ReplyEquipmentSubType;
    serialNumber?: string | null;
    yearInstalled?: number | null;
    existingOrProposed?: number | null;
    mainGasMeter?: boolean | null;
    mainGasMeterNumber?: string | null;
    measureType?: number | null;
    outputRangeMin?: number | null;
    outputRangeMax?: number | null;
    outputUnit?: OutputUnit;
    uncertainty?: number | null;
    comments?: string | null;
    meterFilesList?: Array<ReplyMeterFile> | null;
}