// chpqa-app/src/app/api-services/chpqa-api/generated/model/requestMeter.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OutputUnit } from './outputUnit';


export interface RequestMeter { 
    id?: string | null;
    name: string;
    tagNumber: string;
    measureType: number;
    equipmentTypeId: string;
    equipmentSubTypeId: string;
    serialNumber: string;
    yearInstalled: number;
    existingOrProposed: number;
    mainGasMeter?: boolean | null;
    mainGasMeterNumber?: string | null;
    outputRangeMin: number;
    outputRangeMax: number;
    outputUnit: OutputUnit;
    uncertainty?: number | null;
    comments?: string | null;
}